% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_index_fun_run.R
\name{bootstrap_index_fun_run}
\alias{bootstrap_index_fun_run}
\title{Computes Regional and Coastwide Bootstrapped Relative Abundance Indices}
\usage{
bootstrap_index_fun_run(
  species_vec,
  survey_boundaries,
  ICR_adjust = F,
  data,
  R = 1000,
  return = F,
  ncpus = 1,
  type = "perc",
  plot = T,
  preserve_inter_regional_differences = F
)
}
\arguments{
\item{species_vec}{a character vector of all the species (linked to the variables `N_it` in the dataframe)}

\item{survey_boundaries}{a SpatialPolygonsDataFrame object containing the survey boundary definitions}

\item{ICR_adjust}{if FALSE/TRUE it computes a CPUE-based/instantaneous catch rate-adjusted index}

\item{data}{a SpatialPointsDataFrame object containing the IPHC data}

\item{R}{specifies the number of bootstrap samples desired}

\item{return}{when TRUE it returns the indices}

\item{ncpus}{specifies the number of cores to use in parallel}

\item{type}{specifies the type of bootstrap intervals to use (default percentile)}

\item{plot}{when TRUE it returns the plot.}

\item{preserve_inter_regional_differences}{if TRUE estimated inter-regional differences in mean abundance are shown at a cost of higher variance. Does not affect coastwide index.}
}
\description{
Computes Regional and Coastwide Bootstrapped Relative Abundance Indices
}
