% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored_index_fun.R
\name{censored_index_fun}
\alias{censored_index_fun}
\title{Compute Poisson-lognormal model-based indices of relative abundance}
\usage{
censored_index_fun(
  data,
  survey_boundaries,
  species,
  M = 1000,
  return = T,
  ICR_adjust = F,
  cprop = 1.1,
  nthreads = 1,
  keep = F,
  use_upper_bound = FALSE,
  upper_bound_quantile = 1,
  plot = T,
  allyears = F,
  station_effects = T,
  prior_event = HT.prior(),
  prior_station = HT.prior(),
  init_vals = NULL,
  n_knots = 8,
  seed = 0,
  verbose = F,
  n_trajectories = 10,
  preserve_inter_regional_differences = F
)
}
\arguments{
\item{data}{a SpatialPointsDataFrame object containing the IPHC data}

\item{survey_boundaries}{a SpatialPolygonsDataFrame object containing the survey boundary definitions}

\item{species}{a character name of the species (linked to the variables `N_it` in the dataframe)}

\item{M}{the number of independent Monte Carlo samples from the posterior used to compute indices}

\item{return}{logical stating whether or not to return indices as a data.frame}

\item{ICR_adjust}{A logical determining if the ICR-based scale-factor adjustment should be used}

\item{cprop}{the minimum proportion of baits needed to be removed to induce censorship. If >1 no censorship.}

\item{nthreads}{how many cores do you want to run in parallel when fitting the model?}

\item{keep}{a logical stating if you want to return the INLA model object. Useful for inspecting DIC, properties etc.,}

\item{use_upper_bound}{a logical stating if right-censored or interval-censored response (with concervative upper bound derived using Baranov Catch equation) is desired.}

\item{upper_bound_quantile}{a proportion stating which quantiles of observation to remove censorship from. If greater than 1, do not remove censorship from any observation}

\item{plot}{logical. Do you want to return a ggplot of the indices along with the data.frame?}

\item{allyears}{logical determining if upper bound quantile is computed uniquely for each year (if False) or over all years (if TRUE)}

\item{station_effects}{logical stating if IID station-station random effects wanted}

\item{prior_event}{a prior distribution for the event-event IID random effect standard deviation (see HT.prior for default half t_7(0,1) distribution)}

\item{prior_station}{a prior distribution for the station-station IID random effect standard deviation (see HT.prior for default half t_7(0,1) distribution)}

\item{init_vals}{initial values for fitting model. If NULL let INLA choose them.}

\item{n_knots}{the number of knots used for the temporal 'spline'. Default is 8. More knots means a 'wigglier' temporal effect is allowed, at a risk of higher variance.}

\item{seed}{the seed used for Monte Carlo sampling. Note that 0 means the result is non-reproducible, but the code can be significantly faster.}

\item{verbose}{logical. If TRUE, print INLA diagnostics on the console as the model fits. Can help to diagnose convergence issues.}

\item{n_trajectories}{integer specifying how many Monte Carlo sampled relative abundance indices to plot on a spaghetti plot. This can help with interpreting uncertainty. Suggested value 10.}

\item{preserve_inter_regional_differences}{Logical if TRUE estimated inter-regional differences in mean abundance are shown at a cost of higher variance. Does not affect coastwide index.}
}
\description{
Compute Poisson-lognormal model-based indices of relative abundance
}
