% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coast_Simp.R
\docType{data}
\name{Coast_simp}
\alias{Coast_simp}
\title{Simplifed Coastline of BC, Canada used for building mesh in vignette}
\format{
A SpatialPolygonsDataFrame object:
\describe{
  \item{id}{the index of the polygon}
}
}
\usage{
Coast_simp
}
\description{
A SpatialPolygonsDataFrame object
}
\keyword{datasets}
