% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coastline.R
\docType{data}
\name{Coastline}
\alias{Coastline}
\title{400m - resolution coastline of BC, Canada used for plotting in vignette}
\format{
A SpatialPolygons object:
}
\usage{
Coastline
}
\description{
A SpatialPolygonsDataFrame object
}
\keyword{datasets}
