% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hookCompetition.R
\docType{package}
\name{hookCompetition}
\alias{hookCompetition}
\title{hookCompetition: A package for computing censored-Poisson relative abundance
indices from longline data.}
\description{
The hookCompetition package provides three categories of important functions:
bootstrapped unadjusted methods, non-spatial model-based censored GLMM-based
methods, and spatio-temporal model-based censored ST-GLMM-based methods. The
package is written in INLA which allows for multi-threading.
}
