% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HT.prior.R
\name{HT.prior}
\alias{HT.prior}
\title{The half t7 prior useful for the standard deviation of random effects}
\usage{
HT.prior()
}
\description{
The half t7 prior useful for the standard deviation of random effects
}
