% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\docType{data}
\name{mesh}
\alias{mesh}
\title{An example triangulation mesh of the coastline of BC, Canada used for model fits in vignette}
\format{
A inla.mesh.2d object:
}
\usage{
mesh
}
\description{
A inla.mesh.2d object
}
\keyword{datasets}
