% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Data_hookcomp.R
\name{read_Data_hookcomp}
\alias{read_Data_hookcomp}
\title{Loads the IPHC catch count data for a species. It returns a list containing
two entries:
\code{Reduced_data_sp} is a \code{SpatialPointsDataFrame} containing all the
set-level catch counts of the species specified (e.g. stored as a variable
name `N_it_yelloweye`).
\code{survey_boundaries} are the survey boundaries projected into the same
CRS as the \code{Reduced_data_sp} object.}
\usage{
read_Data_hookcomp(
  species_vec = "yelloweye rockfish",
  at_PBS = F,
  simple_species_vec = "yelloweye",
  data_wd = "./",
  min_dist = 50,
  min_dist_to_boundary = 0,
  years_all_vec = c(1995, 1996, 2003:2012, 2014:2019),
  years_20_vec = c(1997:2002, 2013, 2020),
  survey_boundaries = NULL
)
}
\arguments{
\item{species_vec}{is a character vector of the one or more species desired for modelling}

\item{at_PBS}{is a logical flag determining if the user has access to the gf data. If FALSE, yelloweye rockfish must be the specified species.}

\item{simple_species_vec}{is a character vector containing the desired shortened names of the species}

\item{data_wd}{is a character vector stating the working directory where the rds files will be saved}

\item{min_dist}{is a numeric stating the minimum distance (km) required for a temporary IPHC station to be from a permanent IPHC station to be included}

\item{min_dist_to_boundary}{is a numeric stating the minimum distance (km) required for an IPHC station to be from the survey boundary}

\item{years_all_vec}{is a numeric vector stating the years all hooks were enumerated}

\item{years_20_vec}{is a numeric vector stating the years only 20 hooks were enumerated}

\item{survey_boundaries}{is a sf polygons object containing user-specified survey boundaries for indices to be computed in. If NULL QCS, WCHG, WCVI, HS used. The data.frame *must* contain a factor variable `Region` with levels equal to the names of the regions}
}
\description{
Loads the IPHC catch count data for a species. It returns a list containing
two entries:
\code{Reduced_data_sp} is a \code{SpatialPointsDataFrame} containing all the
set-level catch counts of the species specified (e.g. stored as a variable
name `N_it_yelloweye`).
\code{survey_boundaries} are the survey boundaries projected into the same
CRS as the \code{Reduced_data_sp} object.
}
