% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sets_other_years.R
\docType{data}
\name{sets_other_years}
\alias{sets_other_years}
\title{IPHC data on the survey sets in years 2003 onwards excluding 2013. From gfiphc::get_iphc_sets_info()}
\format{
A tibble object:
\describe{
  \item{E_it}{The effective skate of the set during experiments that enumerate all hooks}
  \item{E_it20}{The effective skate of the set during experiments that enumerate only 20 hooks}
  \item{N_it}{The number of hooks returning bait during experiments that enumerate all hooks}
  \item{N_it20}{The number of hooks returning bait during experiments that enumerate only 20 hooks}
}
}
\usage{
sets_other_years
}
\description{
A tibble object
}
\keyword{datasets}
