% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatio_temporal_plot_fun.R
\name{spatio_temporal_plot_fun}
\alias{spatio_temporal_plot_fun}
\title{Fit spatio-temporal indices using the GMRF-SPDE approximation of lindgren et al 2011}
\usage{
spatio_temporal_plot_fun(
  pred_df_plot,
  year,
  variable,
  hires_COAST,
  plot_figure = T,
  return_figure = F
)
}
\arguments{
\item{pred_df_plot}{the SpatialPixelsDataFrame containing spatiotemporal predictions returned by `spatiotemp_censored_index_fun`}

\item{year}{integer specifying which year to plot}

\item{variable}{character specifying which variable to plot}

\item{hires_COAST}{a SpatialPolygonsDataFrame specifying the coastline used for plotting}

\item{plot_figure}{logical. Do you want to print the plot?}

\item{return_figure}{logical. Do you want to return the plot?}
}
\description{
Fit spatio-temporal indices using the GMRF-SPDE approximation of lindgren et al 2011
}
