% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_boundaries.R
\docType{data}
\name{survey_boundaries}
\alias{survey_boundaries}
\title{SpatialPolygonsDataFrame object containing 4 survey boundaries of BC (HS, QCS, WCHG, WCVI)}
\format{
SpatialPolygonsDataFrame contains:
\describe{
  \item{Region}{Factor variable with levels HS, QCS, WCHG, and WCVI}
  \item{spatio_temporal_CPUE_trajectory}{data.frame containing 10 posterior samples of regional yelloweye relative abundance from the spatio-temporal CPUE-based model}
  \item{spatio_temporal_CPUE_inddf}{data.frame containing posterior summaries of regional yelloweye relative abundance from the spatio-temporal CPUE-based model}
  \item{Other_Objects}{The same as above but for ICR-based and censored models}
}
}
\usage{
survey_boundaries
}
\description{
SpatialPolygonsDataFrame
}
\keyword{datasets}
