% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes_precompiled_data.R
\docType{data}
\name{vignettes_precompiled_data}
\alias{vignettes_precompiled_data}
\title{Precomputed spatio-temporal model fits for the vignettes. This saves time}
\format{
A list:
\describe{
  \item{spatio_temporal_CPUE_plotdf}{SpatialPixelsDataFrame containing mapped posterior quantities from spatio-temporal CPUE-based model}
  \item{spatio_temporal_CPUE_trajectory}{data.frame containing 10 posterior samples of regional yelloweye relative abundance from the spatio-temporal CPUE-based model}
  \item{spatio_temporal_CPUE_inddf}{data.frame containing posterior summaries of regional yelloweye relative abundance from the spatio-temporal CPUE-based model}
  \item{Other_Objects}{The same as above but for ICR-based and censored models}
}
}
\usage{
vignettes_precompiled_data
}
\description{
A list
}
\keyword{datasets}
