% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelloweye_rockfish.R
\docType{data}
\name{yelloweye_rockfish}
\alias{yelloweye_rockfish}
\title{list containing IPHC catch counts of yelloweye rockfish. Obtained from gfiphc::cache_pbs_data_iphc}
\format{
list contains set_counts which is a tibble with variables:
\describe{
  \item{E_it}{The effective skate of the set during experiments that enumerate all hooks}
  \item{E_it20}{The effective skate of the set during experiments that enumerate only 20 hooks}
  \item{N_it}{The number of hooks returning bait during experiments that enumerate all hooks}
  \item{N_it20}{The number of hooks returning bait during experiments that enumerate only 20 hooks}
}
}
\usage{
yelloweye_rockfish
}
\description{
list
}
\keyword{datasets}
